fix build tests
